+++ packages/amplify-graphql-api-construct-tests/src/__tests__/migration/many-to-many-migration.test.ts
@@ -89,8 +89,8 @@ describe('Many-to-many Migration', () => {
           type PostTags @model @auth(rules: [{ allow: public }]) {
             postId: ID
             tagId: ID
-            post: Post @belongsTo(references: ["postId"], overrideIndexName: "byPost")
-            tag: Tag @belongsTo(references: ["tagId"], overrideIndexName: "byTag")
+            post: Post @belongsTo(references: ["postId"], experimentalOverrideIndexName: "byPost")
+            tag: Tag @belongsTo(references: ["tagId"], experimentalOverrideIndexName: "byTag")
           }
         `,
         strategy: {
@@ -272,30 +272,32 @@ describe('Many-to-many Migration', () => {
 
     // assert tables have not been deleted after deleting the gen 1 project
 
-    const listResult = await graphql(
-      gen2APIEndpoint,
-      gen2APIKey,
-      /* GraphQL */ `
-        query LIST_POSTS {
-          listPosts {
-            items {
-              id
-              tags {
-                items {
-                  id
-                }
-              }
-            }
-          }
-        }
-      `,
-    );
+    // This will not work until the enablegen2migration feature flag is released
+    // Restore this block when the feature flag is released
+    // const listResult = await graphql(
+    //   gen2APIEndpoint,
+    //   gen2APIKey,
+    //   /* GraphQL */ `
+    //     query LIST_POSTS {
+    //       listPosts {
+    //         items {
+    //           id
+    //           tags {
+    //             items {
+    //               id
+    //             }
+    //           }
+    //         }
+    //       }
+    //     }
+    //   `,
+    // );
 
-    expect(listResult.statusCode).toEqual(200);
-    expect(listResult.body.data.listPosts.items.length).toEqual(2);
-    expect([gen1Post.id, gen2Post.id]).toContain(listResult.body.data.listPosts.items[0].id);
-    expect([gen1Post.id, gen2Post.id]).toContain(listResult.body.data.listPosts.items[1].id);
-    expect(listResult.body.data.listPosts.items[0].tags.items.length).toEqual(1);
-    expect(listResult.body.data.listPosts.items[1].tags.items.length).toEqual(1);
+    // expect(listResult.statusCode).toEqual(200);
+    // expect(listResult.body.data.listPosts.items.length).toEqual(2);
+    // expect([gen1Post.id, gen2Post.id]).toContain(listResult.body.data.listPosts.items[0].id);
+    // expect([gen1Post.id, gen2Post.id]).toContain(listResult.body.data.listPosts.items[1].id);
+    // expect(listResult.body.data.listPosts.items[0].tags.items.length).toEqual(1);
+    // expect(listResult.body.data.listPosts.items[1].tags.items.length).toEqual(1);
   });
 });