+++ packages/amplify-graphql-relational-transformer/src/__tests__/amplify-graphql-belongs-to-transformer.test.ts
@@ -811,7 +811,7 @@ describe('@belongsTo directive with RDS datasource', () => {
 
       type Comment @model {
         blogId: ID
-        blog: Blog @belongsTo(references: ["blogId"], overrideIndexName: "byBlog")
+        blog: Blog @belongsTo(references: ["blogId"], experimentalOverrideIndexName: "byBlog")
       }
     `;
 
@@ -830,7 +830,7 @@ describe('@belongsTo directive with RDS datasource', () => {
         comment: Comment @hasOne
       }
       type Comment @model {
-        blog: Blog @belongsTo(overrideIndexName: "byBlog")
+        blog: Blog @belongsTo(experimentalOverrideIndexName: "byBlog")
       }
     `;
     expect(() =>
@@ -839,7 +839,7 @@ describe('@belongsTo directive with RDS datasource', () => {
         transformers: [new ModelTransformer(), new HasOneTransformer(), new BelongsToTransformer()],
       }),
     ).toThrowError(
-      'overrideIndexName cannot be used on @belongsTo without references. Modify Comment.blog to use references or remove overrideIndexName.',
+      'experimentalOverrideIndexName cannot be used on @belongsTo without references. Modify Comment.blog to use references or remove experimentalOverrideIndexName.',
     );
   });
 
@@ -851,7 +851,7 @@ describe('@belongsTo directive with RDS datasource', () => {
 
       type Comment @model {
         blogId: ID
-        blog: Blog @belongsTo(references: ["blogId"], overrideIndexName: "byBlog")
+        blog: Blog @belongsTo(references: ["blogId"], experimentalOverrideIndexName: "byBlog")
       }
     `;
 
@@ -870,7 +870,7 @@ describe('@belongsTo directive with RDS datasource', () => {
         comments: [Comment] @hasMany
       }
       type Comment @model {
-        blog: Blog @belongsTo(overrideIndexName: "byBlog")
+        blog: Blog @belongsTo(experimentalOverrideIndexName: "byBlog")
       }
     `;
     expect(() =>
@@ -879,7 +879,7 @@ describe('@belongsTo directive with RDS datasource', () => {
         transformers: [new ModelTransformer(), new HasManyTransformer(), new BelongsToTransformer()],
       }),
     ).toThrowError(
-      'overrideIndexName cannot be used on @belongsTo without references. Modify Comment.blog to use references or remove overrideIndexName.',
+      'experimentalOverrideIndexName cannot be used on @belongsTo without references. Modify Comment.blog to use references or remove experimentalOverrideIndexName.',
     );
   });
 });