+++ packages/amplify-graphql-relational-transformer/src/has-many/has-many-directive-ddb-references-transformer.ts
@@ -14,7 +14,7 @@ import {
 import { HasManyDirectiveConfiguration } from '../types';
 import {
   ensureReferencesArray,
-  getOverrideIndexName,
+  getexperimentalOverrideIndexName,
   getReferencesNodes,
   registerHasManyForeignKeyMappings,
   validateParentReferencesFields,
@@ -73,9 +73,9 @@ export class HasManyDirectiveDDBReferencesTransformer implements DataSourceBased
     const fieldName = config.field.name.value;
     config.referenceNodes = getReferencesNodes(config, context);
     validateReferencesBidirectionality(config);
-    const overrideIndexName = getOverrideIndexName(config);
-    if (overrideIndexName) {
-      config.indexName = overrideIndexName;
+    const experimentalOverrideIndexName = getexperimentalOverrideIndexName(config);
+    if (experimentalOverrideIndexName) {
+      config.indexName = experimentalOverrideIndexName;
     } else {
       config.indexName = `gsi-${objectName}.${fieldName}`;
     }