+++ packages/amplify-graphql-relational-transformer/src/types.ts
@@ -66,7 +66,7 @@ export type BelongsToDirectiveConfiguration = {
   relationType: 'hasOne' | 'hasMany';
   relatedTypeIndex: FieldDefinitionNode[];
   connectionFields: string[];
-  overrideIndexName: string;
+  experimentalOverrideIndexName: string;
 };
 
 export type ManyToManyDirectiveConfiguration = {